Removed @pigi/operator in favor of singular @pigi/core
