+++ packages/core/package.json
@@ -1,28 +1,52 @@
 {
   "name": "@pigi/core",
   "version": "0.0.2-alpha.1",
-  "description": "Plasma Group client",
-  "main": "build/index.js",
-  "files": [
-    "build/**/.js"
-  ],
-  "scripts": {
-  },
+  "description": "Plasma Group generalized plasma client",
   "keywords": [
     "plasma",
     "group",
     "ethereum",
     "client"
   ],
-  "author": "Plasma Group",
+  "homepage": "https://github.com/plasma-group/pigi/tree/master/packages/core#readme",
+  "bugs": "https://github.com/plasma-group/pigi/labels/%40pigi%2Fcore",
   "license": "MIT",
+  "author": "Plasma Group",
+  "files": [
+    "build/**/.js"
+  ],
+  "main": "build/index.js",
   "repository": {
     "type": "git",
     "url": "https://github.com/plasma-group/pigi.git"
   },
+  "scripts": {
+    "test": "mocha --require ts-node/register 'test/**/*.spec.ts'",
+    "lint": "tslint --format stylish --project .",
+    "fix": "prettier --config ./.prettierrc.js --write 'index.ts' '{src,test}/**/*.ts'",
+    "build": "tsc -p ."
+  },
   "dependencies": {
+    "abstract-leveldown": "^6.0.3",
+    "axios": "^0.19.0",
+    "bn.js": "^4.11.8",
+    "body-parser": "^1.19.0",
+    "debug": "^4.1.1",
+    "eth-lib": "^0.2.8",
+    "express": "^4.17.1",
+    "uuid": "^3.3.2",
+    "web3": "^1.0.0-beta.55",
+    "web3-utils": "^1.0.0-beta.55"
   },
   "devDependencies": {
+    "@types/chai": "^4.1.7",
+    "@types/mocha": "^5.2.7",
+    "@types/node": "^12.0.7",
+    "chai": "^4.2.0",
+    "chai-as-promised": "^7.1.1",
+    "mocha": "^6.1.4",
+    "ts-node": "^8.2.0",
+    "typescript": "^3.5.1"
   },
   "publishConfig": {
     "access": "public"
