+++ packages/core/src/app/common/db/index.ts
@@ -2,3 +2,4 @@ export * from './key/key'
 export * from './bucket'
 export * from './db'
 export * from './iterator'
+export * from './range-db'
