+++ packages/core/src/app/common/db/range-db.ts
@@ -9,7 +9,7 @@ import {
   RangeStore,
   RangeEntry,
   Endianness,
-} from '../../interfaces/db/range-db.interface'
+} from '../../../interfaces/common/db/range-db.interface'
 
 /* Logging */
 import debug from 'debug'
