+++ packages/core/src/app/common/utils/index.ts
@@ -3,3 +3,4 @@ export * from './debug-logger'
 export * from './misc'
 export * from './range-store'
 export * from './type-guards'
+export * from './operator-utils'
