+++ packages/core/src/app/common/utils/operator-utils.ts
