+++ packages/core/src/interfaces/common/db/index.ts
@@ -1,2 +1,3 @@
 export * from './db.interface'
 export * from './key.interface'
+export * from './range-db.interface'
