+++ packages/core/src/interfaces/common/db/range-db.interface.ts
