--- packages/operator/.prettierrc.js
@@ -1,1 +1,0 @@
-module.exports = require('@pigi/prettier-config')
