--- packages/operator/index.ts
@@ -1,5 +1,0 @@
-import path = require('path')
-const rootPath = __dirname
-const dbRootPath = path.join(__dirname, 'db')
-
-export { rootPath, dbRootPath }
