--- packages/operator/src/app/constants/db-defaults.ts
@@ -1,3 +1,0 @@
-export const ETH_DB_FILENAME = 'eth-config.json'
-export const TEST_DB_DIR = './operator-db-test/'
-export const KEYSTORE_DIR = 'operator-keystore'
