--- packages/operator/src/app/constants/index.ts
@@ -1,4 +1,0 @@
-export * from './db-defaults'
-export * from './db-prefixes'
-export * from './rpc-methods'
-export * from './serialization'
