--- packages/operator/src/app/constants/rpc-methods.ts
@@ -1,16 +1,0 @@
-export const INIT_METHOD = 'init'
-export const GET_ETH_INFO_METHOD = 'getEthInfo'
-export const STATE_METHODS = {
-  DEPOSIT_METHOD: 'deposit',
-  ADD_TX_METHOD: 'addTransaction',
-  NEW_BLOCK_METHOD: 'newBlock',
-  GET_BLOCK_NUMBER_METHOD: 'getBlockNumber',
-  GET_TXS_METHOD: 'getTransactions',
-  GET_RECENT_TXS_METHOD: 'getRecentTransactions',
-}
-export const BLOCK_METHODS = {
-  GET_HISTORY_PROOF: 'getHistoryProof',
-  GET_BLOCK_METADATA_METHOD: 'getBlockMetadata',
-  GET_TX_FROM_HASH_METHOD: 'getTxFromHash',
-  GET_BLOCK_TXS_METHOD: 'getBlockTransactions',
-}
