--- packages/operator/src/app/exceptions/index.ts
@@ -1,1 +1,0 @@
-export * from './server-already-started.exception'
