--- packages/operator/src/app/exceptions/messages.ts
@@ -1,1 +1,0 @@
-export const SERVER_ALREADY_STARTED_MESSAGE = `RPC server has already been started.`
